Xib working
