--- John Hearn/Gitcha/Gitcha/Models/RepositoryCell.swift
@@ -1,29 +1,0 @@
-//
-//  RepositoryCell.swift
-//  Gitcha
-//
-//  Created by John D Hearn on 11/1/16.
-//  Copyright © 2016 Bastardized Productions. All rights reserved.
-//
-
-import UIKit
-
-class RepositoryCell: UITableViewCell {
-
-    override func prepareForReuse() {
-        super.prepareForReuse()
-        self.textLabel?.text = ""
-        self.detailTextLabel?.text = ""
-    }
-
-    override func awakeFromNib() {
-        super.awakeFromNib()
-    }
-
-    override func setSelected(_ selected: Bool, animated: Bool) {
-        super.setSelected(selected, animated: animated)
-
-        // Configure the view for the selected state
-    }
-
-}
