+++ John Hearn/Gitcha/Gitcha/Views/RepositoryTableViewCell.xib
@@ -11,7 +11,7 @@
     <objects>
         <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
         <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <tableViewCell clipsSubviews="YES" contentMode="scaleToFill" selectionStyle="default" indentationWidth="10" reuseIdentifier="RepositoryCell" rowHeight="93" id="oFk-gs-ufw">
+        <tableViewCell clipsSubviews="YES" contentMode="scaleToFill" selectionStyle="default" indentationWidth="10" reuseIdentifier="RepositoryTableViewCell" rowHeight="93" id="oFk-gs-ufw" customClass="RepositoryTableViewCell" customModule="Gitcha" customModuleProvider="target">
             <rect key="frame" x="0.0" y="0.0" width="375" height="93"/>
             <autoresizingMask key="autoresizingMask"/>
             <tableViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" tableViewCell="oFk-gs-ufw" id="3DT-0V-Wex">
@@ -36,6 +36,7 @@
                 </subviews>
                 <constraints>
                     <constraint firstAttribute="trailing" secondItem="D8Y-9a-dg8" secondAttribute="trailing" constant="8" id="3Fy-ud-cAY"/>
+                    <constraint firstAttribute="bottom" relation="greaterThanOrEqual" secondItem="XIt-fV-A6u" secondAttribute="bottom" constant="8" id="4wB-fQ-34l"/>
                     <constraint firstItem="D8Y-9a-dg8" firstAttribute="leading" secondItem="3DT-0V-Wex" secondAttribute="leading" constant="8" id="FfK-7C-Fcb"/>
                     <constraint firstItem="D8Y-9a-dg8" firstAttribute="top" secondItem="3DT-0V-Wex" secondAttribute="top" constant="8" id="PCF-nQ-mrL"/>
                     <constraint firstItem="XIt-fV-A6u" firstAttribute="leading" secondItem="3DT-0V-Wex" secondAttribute="leading" constant="8" id="VxT-xx-ITp"/>
@@ -43,6 +44,10 @@
                     <constraint firstItem="XIt-fV-A6u" firstAttribute="top" secondItem="D8Y-9a-dg8" secondAttribute="bottom" constant="8" id="viu-iX-LvS"/>
                 </constraints>
             </tableViewCellContentView>
+            <connections>
+                <outlet property="descriptionLabel" destination="XIt-fV-A6u" id="B9m-Hf-Lc1"/>
+                <outlet property="repositoryLabel" destination="D8Y-9a-dg8" id="dor-5D-4ze"/>
+            </connections>
             <point key="canvasLocation" x="33.5" y="78.5"/>
         </tableViewCell>
     </objects>
